Move character-marker menu generator out of the extension-host graph
generateCharacterMarkerMenuListItems needs a lucide-react icon
(RemoveFormatting) for its remove row, and it lived in
platform-scripture-editor.utils.ts, which main.ts imports 12 values from. The
icon is a runtime value, so TypeScript could not elide it and React entered the
extension-host module graph. Extension builds externalize the app's
dependencies, so the bundle emitted a runtime require('react'), which the
extension-host import shim rejects:

  Extension 'platformScriptureEditor' threw while activating!
  Requiring other than papi is not allowed in extensions!
  Rejected require('react').

The whole extension failed to activate, so no scripture editor would open.

Move the generator, CHARACTER_MARKER_MENU_STRING_KEYS and
REMOVE_CHARACTER_MARKER_KEY into a new character-marker-menu.utils.ts that only
the web view imports, and move their tests alongside. Web views are exempt from
the restriction because main.ts imports them as ?inline strings, so they form a
separate graph where React is legal. No signature, behavior or TSDoc change.

This is a boundary, not a workaround: platform-scripture-editor.utils.ts is
imported by both the extension host and the web view, so it cannot hold UI
values. Note that nothing enforces this today - a future icon or JSX import into
a main-reachable module breaks activation the same way, and it fails at runtime
rather than at build or lint time.

Verified: typecheck clean, lint clean, 115 tests pass, and
"finished activating platformScriptureEditor" in the log where it previously
threw.
